requirements.txt
kaggle==1.6.17
pandas==2.2.2
google-cloud-storage==2.16.0
pyarrow==16.0.0
dbt-bigquery
streamlit
plotly==5.22.0
google-cloud-bigquery==3.21.0
google-cloud-bigquery-storage==2.25.0
db-dtypes==1.2.0